todoapp_django/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.2
djangorestframework==3.14.0
filelock==3.13.1
httpie==3.2.2
identify==2.5.35
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.6.2
Pygments==2.17.2
PySocks==1.7.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
setuptools==69.1.1
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1